Setting the stage for LocVicinity listener
Thinking of setting optional flags for rendezes, setting ranges for whenever sender A sends a rendez, and when A is in the vicinity of the location.  Settings should be 1 to 1, meaning that if Sender A has the range set, he/she sends notification, and not mutually unless Reciever B also has their own settings set on a range.

Flags params:
0 - 100 meters?
1- 1 mile?
2 - 3-5 miles?

Need to set a new table with the relational scheme set to the rendez id, as well as columns set to flags and times.  I don't think location data needs to be sent since the location of the rendez is already known to both parties, so by just using the flag A and B will know the general vicinity that the person pinged their location.

rendezChatInRangeViewController:

Process: Background running
Listing in order of most recently made/sent rendezes between parties (1 to 1/1 to many/group)

Reqs: Listener in the AppDelegate that keeps tabs on all flags that have been set for each